week8-getting-started-with-aws-strands-agent/getting-started-strands-agents/requirements.txt
strands-agents[litellm]
strands-agents-tools